requirements.txt
colorama
requests
estrapy-api
aiohoripy
Flask==2.0.1
discord.py
discord.py[voice]
yt-dlp
youtube_search==2.1.2
youtube-search-python==1.6.6
python-dotenv==1.0.0
googlesearch-python==1.2.3
aiohttp==3.8.6
Random-Word
httpx
holodex
hmtai
Pillow
Ffmpeg
PyNaCl
youtube-search-python
youtube_dl
jokeapi
googletrans
captcha
qrcode[pil]
googletrans
akanekopy
langdetect
langid
translate
random-dog
anime-api
waifu-py
colorama
pokeapi
roman
nltk
fuzzywuzzy
vaderSentiment
inflect
spacy
emoji
pytube
pillow
praw
beautifulsoup4
fandom-py
pokebase
faker
Faker
python-Levenshtein
discord-webhook
psutil
